Make basic_json allocate-aware, as mentioned in #161
This is an attempt to make it work with #164.
The thing is that the current code attempts to define containers with incomplete types,
this is not supported by the standard and does not work unordered_map. std::vector and
std::map work but we might be in the realm of undefined behavior.