continuing work on our own strtold - reduced to 3 failing assertions
this changes a few things. the name of the function to better represent
what it does (string to json number) - we aren't strtoul compliant so
best not to lie about that.

we remove template code and just use one function. this makes it a bit
simpler and we can always add it back if it buys us anything. the
penalty for casting down from long double seems to be inconsequential.