Fix include brackets so they correspond to local includes instead of system includes
When building code relying on this library in Xcode, the build process throws an error (not a warning!) stating its inability to locate the presumed system includes. The `<>` angle brackets on `#include` indicate a system-level include, not a user-level include. By swapping to quotation marks for the `#include` we more accurately represent the content of the included files and Xcode builds without complaint!