Give message about the version of the code used only once
When consuming the library via add_subdirectory() using CMake, the
message about the version of the code being used was given every time
CMake was re-run, even if it happened due to the changes in the super
project completely unrelated to this library.

This was somewhat annoying, so avoid this by remembering the version of
the code used in the cache and not giving the message about it again if
it didn't change since the last run.

Signed-off-by: Vadim Zeitlin <vz-github@zeitlins.org>